[ANCHOR-955] Add SEP-45 config validation (#1639)
### Description

This PR adds SEP-45 configuration validation. I decided to do the secret
validation in the secret config implementation. I think it's the best
place to implement it. Dependent configs, like SEP-45, should just check
that the relevant properties are set. We should refactor the rest of the
configuration to follow the same pattern.

### Context

It's currently missing.

### Testing

- `./gradlew test`

### Documentation

N/A

### Known limitations

N/A