Enable event processing for callback API (#1662)
### Description

This enables the event processor to send events to the business server
in the local Kubernetes setup. SEP-6 would not work without event
processing enabled.

### Context

This config was true by default in v2 but changed to false in v3. The
helm chart values were never updated.

### Testing

- `./gradlew test`
- Tested that it works using the demo wallet.

### Documentation

N/A

### Known limitations

N/A