Fix: Allow ref server chart to configure `rpcEnabled` (#1875)
### Description

This change allows users to configure `rpcEnabled` in the reference
server chart.

### Context

It's currently hardcoded to false in the config map. We are currently
only allowing `rpcEnabled` to be set, as this is a hotfix for the
testanchor. The other configurations will be made configurable in
another PR that will be released in a future version.

### Testing

- `./gradlew test`

### Documentation

N/A

### Known limitations

N/A